Always complete the old auto-upgrade procedure
This is required for clients upgrading from versions <=1.2.7
